+++ Cmdr/Shared/Argument.lua
@@ -1,5 +1,13 @@
 local Util = require(script.Parent.Util)
 
+local function unescapeOperators(text)
+	for _, operator in ipairs({"%.", "%?", "%*", "%*%*"}) do
+		text = text:gsub("\\%" .. operator, operator:gsub("%%", ""))
+	end
+
+	return text
+end
+
 local Argument = {}
 Argument.__index = Argument
 
@@ -16,6 +24,8 @@ function Argument.new (command, argumentObject, value)
 		RawSegments = {}; -- The raw, unparsed segments (if the raw value was comma-sep)
 		TransformedValues = {}; -- The transformed value (generated later)
 		Prefix = ""; -- The prefix for this command (%Team)
+		TextSegmentInProgress = ""; -- The text of the raw segment the user is currently typing.
+		RawSegmentsAreAutocomplete = false;
 	}
 
 	if type(argumentObject.Type) == "table" then
@@ -42,6 +52,17 @@ function Argument.new (command, argumentObject, value)
 	return self
 end
 
+function Argument:GetDefaultAutocomplete()
+	if self.Type.Autocomplete then
+		local transformedEmpty = self:TransformSegment("")
+		local strings, options = self.Type.Autocomplete(transformedEmpty)
+
+		return strings, options or {}
+	end
+
+	return {}
+end
+
 --- Calls the transform function on this argument.
 -- The return value(s) from this function are passed to all of the other argument methods.
 -- Called automatically at instantiation
@@ -50,10 +71,72 @@ function Argument:Transform()
 		return
 	end
 
-	if self.Type.Listable and #self.RawValue > 0 then
-		local rawSegments = Util.SplitStringSimple(self.RawValue, ",")
+	local rawValue = self.RawValue
 
-		if self.RawValue:sub(#self.RawValue, #self.RawValue) == "," then
+	if rawValue == "." and self.Type.Default then
+		rawValue = self.Type.Default(self.Executor) or ""
+		self.RawSegmentsAreAutocomplete = true
+	end
+
+	if rawValue == "?" and self.Type.Autocomplete then
+		local strings, options = self:GetDefaultAutocomplete()
+
+		if not options.IsPartial and #strings > 0 then
+			rawValue = strings[math.random(1, #strings)]
+			self.RawSegmentsAreAutocomplete = true
+		end
+
+	end
+
+	if self.Type.Listable and #self.RawValue > 0 then
+		local randomMatch = rawValue:match("^%?(%d+)$")
+		if randomMatch then
+			local maxSize = tonumber(randomMatch)
+
+			if maxSize and maxSize > 0 then
+				local items = {}
+				local remainingItems, options = self:GetDefaultAutocomplete()
+
+				if not options.IsPartial and #remainingItems > 0 then
+					for _ = 1, math.min(maxSize, #remainingItems) do
+						table.insert(items, table.remove(remainingItems, math.random(1, #remainingItems)))
+					end
+
+					rawValue = table.concat(items, ",")
+					self.RawSegmentsAreAutocomplete = true
+				end
+			end
+		elseif rawValue == "*" or rawValue == "**" then
+			local strings, options = self:GetDefaultAutocomplete()
+
+			if not options.IsPartial and #strings > 0 then
+				if rawValue == "**" and self.Type.Default then
+					local defaultString = self.Type.Default(self.Executor) or ""
+
+					for i, string in ipairs(strings) do
+						if string == defaultString then
+							table.remove(strings, i)
+						end
+					end
+				end
+
+				rawValue = table.concat(
+					strings,
+					","
+				)
+				self.RawSegmentsAreAutocomplete = true
+			end
+		end
+
+		rawValue = unescapeOperators(rawValue)
+
+		local rawSegments = Util.SplitStringSimple(rawValue, ",")
+
+		if #rawSegments == 0 then
+			rawSegments = {""}
+		end
+
+		if rawValue:sub(#rawValue, #rawValue) == "," then
 			rawSegments[#rawSegments + 1] = "" -- makes auto complete tick over right after pressing ,
 		end
 
@@ -61,9 +144,14 @@ function Argument:Transform()
 			self.RawSegments[i] = rawSegment
 			self.TransformedValues[i] = { self:TransformSegment(rawSegment) }
 		end
+
+		self.TextSegmentInProgress = Util.SplitStringSimple(self.RawValue, ",")[1]
 	else
-		self.RawSegments[1] = self.RawValue
-		self.TransformedValues[1] = { self:TransformSegment(self.RawValue) }
+		rawValue = unescapeOperators(rawValue)
+
+		self.RawSegments[1] = unescapeOperators(rawValue)
+		self.TransformedValues[1] = { self:TransformSegment(rawValue) }
+		self.TextSegmentInProgress = self.RawValue
 	end
 end
 
@@ -84,6 +172,10 @@ end
 function Argument:Validate(isFinal)
 	if self.RawValue == nil or #self.RawValue == 0 and self.Required == false then
 		return true
+	end
+
+	if self.Required and (self.RawSegments[1] == nil or #self.RawSegments[1] == 0) then
+		return false, "This argument is required."
 	end
 
 	if self.Type.Validate or self.Type.ValidateOnce then